+++ openqemist/tests/electronic_structure_solvers/vqe_solver/test_vqe_solver_microsoft_qsharp.py
@@ -1,5 +1,5 @@
 #   Copyright 2019 1QBit
-#   
+#
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
 #   You may obtain a copy of the License at
@@ -18,7 +18,6 @@ from enum import Enum
 from pyscf import gto, scf
 
 from openqemist.electronic_structure_solvers import VQESolver, FCISolver
-from openqemist.quantum_solvers.parametric_quantum_solver import ParametricQuantumSolver
 
 H2 = """
    H 0.00 0.00 0.0
@@ -29,6 +28,7 @@ H2 = """
 class VQESolverTest(unittest.TestCase):
 
     def test_h2_sto3g(self):
+        """ Test the converged energye of VQE """
         from openqemist.quantum_solvers import MicrosoftQSharpParametricSolver
 
         mol = gto.Mole()
@@ -44,11 +44,10 @@ class VQESolverTest(unittest.TestCase):
         #solver.optimizer =
 
         energy = solver.simulate(mol)
+        self.assertAlmostEqual(energy, -1.1372704178510415, delta=1e-3)
 
-        self.assertAlmostEqual(energy, -1.1372704178510415, places=3)
-
-    @unittest.skip("After a sign change to the t2 amplitudes, may be quick enough to converge for unittests")
     def test_h2_321g(self):
+        """ Test the converged energye of VQE """
         from openqemist.quantum_solvers import MicrosoftQSharpParametricSolver
 
         mol = gto.Mole()
@@ -61,11 +60,12 @@ class VQESolverTest(unittest.TestCase):
         solver = VQESolver()
         solver.hardware_backend_type = MicrosoftQSharpParametricSolver
         solver.ansatz_type = MicrosoftQSharpParametricSolver.Ansatze.UCCSD
-        #solver.optimizer =
 
+        solver.initial_var_params = [-0.003013364325590502, 0.005694780930328863, 0.00035386539723315964,
+                                     0.04244748156400211, 0.02380828037656643, 0.018789486618923344,
+                                     0.0007322518664444671, 0.04376425859874745, 8.503425743572844e-05]
         energy = solver.simulate(mol)
-
-        self.assertAlmostEqual(energy, -1.1478300615818977, places=3)
+        self.assertAlmostEqual(energy, -1.1478300615818977, delta=1e-3)
 
 
 if __name__ == "__main__":